+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 * Revert added alias to_h for to_hash ([#277](https://github.com/railsconfig/config/issues/277))
 
+### Changes
+
+* Raise explicit error on environment variable conflicts ([#293](https://github.com/railsconfig/config/issues/293))
+
 ## 2.2.2
 
 ### Bug fixes